backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
motor
pymongo
strands-agents
python-dotenv
